agzamov/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agzamov"
version = "0.1.0"
description = "Agzamov Test — game-theoretic benchmark for AI memory infrastructure"
readme = "../agzamov-test.md"
requires-python = ">=3.12"
license = "Apache-2.0"
authors = [
    { name = "Ali Agzamov", email = "ali.agzamov@gmail.com" },
]
dependencies = [
    "python-chess>=1.10",
    "anthropic>=0.40",
    "stockfish>=3.28",
    "scipy>=1.14",
    "numpy>=2.0",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.0",
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
]

[project.scripts]
agzamov = "agzamov.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["."]
